Created initial tests for findi_scan.py
Created the first two tests to ensure that findi_scan.py functions properly.

The first test scans an IP address to ensure that it's open on port 80 and checks to see that the webpage at this IP address is blank.

The second test scans an IP address to ensure that it's open on port 80 and checks to see that the webpage at this IP address contains data.

Created an additional method for getting the page size which was needed for testing, removed more print statements, and fixed a checkstyle issue

This commit address issue #2