Added Coveralls support
Added Coveralls support to the program by adding pytest-cov to the requirements.txt and using it in the script portion of .travis.yml. This will allow for code coverage to be generated upon each build of the program and will pass this information to Coveralls for processing.

This commit addresses issue #2