Switched from Zerorpc to Python Shell for IPC
After trying to update zerorpc to be used with the current version of Electron unsuccessfully, I decided to utilize python shell to serve as the IPC mechanism to allow my python scripts to interact with the electron front end.

Removed old code that utliized zerorpc. Created a wrapper class for the findi_scan script and a scanner js class to handle connecting the front end and backend together.

This commit addresses issue #3