Fixed infinite redirect issue
Fixed an issue where the script will run indefinitely due to never incrementing the redirect counter variable. This issue arose as a result of me breaking the script up into multiple methods without knowing why the script was structured in the way it was. The redirect counter was created in the main thread method and when I was incrementing it in the submethod, it wouldn't get incremented due to being out of scope. To fix this issue, I added an if statement that would increment the redirect counter in the main thread method if tryAgain remained true, rather than incrementing the redirect counter in the submethod.

This commit addresses issue #8