Added a lock to findi_scan.py
Added a lock to the scanning script to ensure that multiple threads don't attempt to change the value of the global running counter variable at the same time. Also added the same code to prevent the same thing from happening to the closed counter variable as well. This commit addresses issue #7