Update findi_list.py
Fixed an issue where an error in opening or loading an info.txt file for an IP would cause the script to stop running. This prevented all following items in the openIPs folder from being printed out.

Cleaned up the formatting in the document and the print statements

This commit addresses issue #1